requirements.txt
certifi==2020.12.5
click==7.1.2
cycler==0.10.0
Flask==1.1.2
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.11.3
Keras==2.4.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
numpy==1.19.5
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
scipy==1.5.4
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1
gunicorn==19.0.0
tensorflow-cpu==2.4.1
opencv-python==4.5.1.48
